backend/requirements.txt
Flask==2.3.3
yt-dlp>=2023.10.13
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0